unicafe-project/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
firebase-admin==6.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.0